Get location and subscriptionID from IMDS when useInstanceMetadata is true
